requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.9
chromadb==0.5.0
streamlit
pandas==2.0.2
python-dotenv==1.0.0
selenium==4.21.0
bs4
sentence-transformers
tenacity
# pysqlite3-binary
PyPDF2
python-docx
